Adding support for the Twitter streaming API
If we instantiate the bot with the new --stream flag, it will
create a version of the bot that listens to the streaming API for events on
our feed. If it detects a quoted tweet, it writes a file to tell the regular
cron-job instance of the bot to fave that tweet for us.